Move ConditionsHolder and Expression to own folder
